requirements.txt
django
graphene
graphene-django
PyJWT
graphql-core